+++ src/better_trace/better_trace.py
@@ -1,4 +1,5 @@
-from types import TracebackType, ModuleType
+from types import TracebackType
+from typing import Any
 from dataclasses import dataclass
 from warnings import warn
 import sys
@@ -95,7 +96,7 @@ def _show_context(filename: str, lineno: int, context: int = 2):
         for i in range(start, end + 1):
             line = linecache.getline(filename, i).rstrip("\n")
             prefix = "❱ " if i == lineno else "  "
-            print(f"{prefix}{i:4} | {line}")
+            print(f"{prefix}{i:4} │ {line}")
         return
 
     if os.path.exists(filename):
@@ -129,6 +130,42 @@ def _show_context(filename: str, lineno: int, context: int = 2):
                 line_numbers=True,
             )
         )
+
+
+def _give_filtered_locals(tb: TracebackType) -> list[Any]:
+    filtered = []
+    for k, v in tb.tb_frame.f_locals.items():
+        if k.lower() in [
+            "password",
+            "token",
+            "key",
+            "api_key",
+            "api_token",
+            "api_password",
+            "secret",
+            "secrets",
+            "passwords",
+            "tokens",
+            "keys",
+            "api_passwords",
+            "api_tokens",
+            "api_keys",
+        ]:
+            continue
+        if k.startswith("__"):
+            continue
+        if inspect.ismodule(v):
+            continue
+        if inspect.isfunction(v):
+            continue
+        if isinstance(v, BaseException):
+            continue
+        try:
+            val = _safe_repr.repr(v)
+        except Exception:
+            val = "<repr broken>"
+        filtered.append((k, val, type(v).__name__))
+    return filtered
 
 
 def _initialize_mode(mode: str) -> str:
@@ -196,10 +233,10 @@ def _print_exception_group(exc: ExceptionGroup, level: int = 0, index_prefix: st
 
     title = f"{index_prefix}" if index_prefix else "Exception Group"
     if not _has_rich:
-        print(f"{indent}{title.center(50, '-')}")
+        print(f"{indent}{title.center(50, '─')}")
         print(f"Message: {str(exc) or '<no message available>'}")
     else:
-        print(f"{indent}[red]{title.center(50, '-')}[/red]")
+        print(f"{indent}[red]{title.center(50, '─')}[/red]")
         print(
             f"[cyan][bold]Message[/bold]: {str(exc) or '<no message available>'}[/cyan]"
         )
@@ -235,18 +272,184 @@ def _render_syntax_error(exc: SyntaxError, heading: bool = True) -> None:
 
     if not _has_rich:
         if heading:
-            print(f"Parsing Error".center(50, "-"))
+            print(f"Parsing Error".center(50, "─"))
         print(f'File "{exc.filename}", line {exc.lineno}')
         print(f"  {line}")
         print(f'  {" " * start + "^" * max(1, end - start)}')
         print(f"{type(exc).__name__}: {exc.msg}")
     else:
         if heading:
-            print(f"[red]Parsing Error[/red]".center(50, "-"))
+            print(f"[red]Parsing Error[/red]".center(50, "─"))
         print(f'File "{exc.filename}", line {exc.lineno}')
         print(f"[red]  {line}[/red]")
         print(f'[red]  {" " * start + "^" * max(1, end - start)}[/red]')
         print(f"[red][bold]{type(exc).__name__}[/bold]: {exc.msg}[/red]")
+
+
+def _print_debug(
+    title: str,
+    frames: traceback.StackSummary,
+    exc_type: type[BaseException],
+    exc: BaseException,
+    tb: TracebackType,
+) -> None:
+
+    if not _has_rich:
+        print(f"{title + ' (Debug mode)'}".center(50, "─") if title is not None else "")
+    else:
+        print(
+            f"[red]{title + ' (Debug mode)'}[/red]".center(50, "─")
+            if title is not None
+            else ""
+        )
+    prev_key = None
+    count = 0
+    prev_frame = None
+    last_tb = tb
+    while last_tb.tb_next is not None:
+        last_tb = last_tb.tb_next
+
+    for frame in frames:
+        key = (frame.filename, frame.name)
+        if key == prev_key:
+            count += 1
+        else:
+            if prev_frame:
+                if not _has_rich:
+                    print(
+                        f'File "{prev_frame.filename}", line {prev_frame.lineno}, in {prev_frame.name}'
+                    )
+                    if not prev_frame.line:
+                        print("  <line unavailable>  ")
+                    else:
+                        _show_context(prev_frame.filename, prev_frame.lineno)
+                    if count > 3:
+                        print(f"(Previous line repeated {count-1} times)")
+                else:
+                    print(
+                        f'File "{prev_frame.filename}", line {prev_frame.lineno}, in [yellow][bold]{prev_frame.name}[/bold][/yellow]'
+                    )
+                    if not prev_frame.line:
+                        print("[red bold]  <line unavailable> [/red bold]")
+                    else:
+                        _show_context(prev_frame.filename, prev_frame.lineno)
+                    if count > 3:
+                        print(
+                            f"[cyan](Previous line repeated {count-1} more times)[/cyan]"
+                        )
+                filtered = _give_filtered_locals(tb)
+                if config.show_locals and filtered:
+                    if not _has_rich:
+                        print("Local variables:")
+                    else:
+                        print("[yellow]Local variables:[/yellow]")
+                for k, v, t in filtered:
+                    print(f" {k} ({t}) = {v}")
+
+                print("─" * 40)
+                tb = tb.tb_next
+            prev_frame = frame
+            prev_key = key
+            count = 1
+    if prev_frame:
+        if not _has_rich:
+            print(
+                f'File "{prev_frame.filename}", line {prev_frame.lineno}, in {prev_frame.name}'
+            )
+            if not prev_frame.line:
+                print("  <line unavailable>  ")
+            else:
+                _show_context(prev_frame.filename, prev_frame.lineno)
+            if count > 3:
+                print(f"(Previous line repeated {count-1} times)")
+        else:
+            print(
+                f'File "{prev_frame.filename}", line {prev_frame.lineno}, in [yellow][bold]{prev_frame.name}[/bold][/yellow]'
+            )
+            if not prev_frame.line:
+                print("[red bold]  <line unavailable> [/red bold]")
+            else:
+                _show_context(prev_frame.filename, prev_frame.lineno)
+            if count > 3:
+                print(f"[cyan](Previous line repeated {count-1} more times)[/cyan]")
+            filtered = _give_filtered_locals(tb)
+            if config.show_locals and filtered:
+                if not _has_rich:
+                    print("Local variables:")
+                else:
+                    print("[yellow]Local variables:[/yellow]")
+            for k, v, t in filtered:
+                print(f" {k} ({t}) = {v}")
+        print("─" * 40)
+        tb = tb.tb_next
+    if _has_rich:
+        print("[yellow]Call stack:[/yellow]")
+    else:
+        print("Call stack:")
+
+    indentation = 3
+    frames_length = len(frames)
+    for frame in frames:
+        if _has_rich:
+            if frames_length != 1:
+                print(f"{' ' * indentation} └─ [cyan][bold]{frame.name}[/bold][/cyan]")
+                indentation += 3
+                frames_length -= 1
+            else:
+                print(f"{' ' * indentation} └─ ", end="")
+                console.print(
+                    Syntax(
+                        frame.line,
+                        "python",
+                        theme=config.theme,
+                        background_color=config.background_color,
+                    )
+                )
+        else:
+            if frames_length != 1:
+                print(f"{' ' * indentation} └─ {frame.name}")
+                indentation += 3
+                frames_length -= 1
+            else:
+                print(f"{' ' * indentation} └─ {frame.line}")
+    print("─" * 40)
+
+    name = exc_type.__name__ or "UnknownError"
+    msg = str(exc) or "<no message provided>"
+    issyntaxerror = isinstance(exc, SyntaxError)
+
+    if not issyntaxerror:
+        has_filename = False
+        has_len = False
+    elif (exc.filename is None) or (exc.lineno is None):
+        has_filename = False
+        has_len = False
+    else:
+        has_filename = True
+        has_len = True
+
+    if not (issyntaxerror and has_filename and has_len):
+        if not _has_rich:
+            print(f"{name}: {msg}")
+        else:
+            print(f"[red][bold]{name}[/bold][/red]: [red]{msg}[/red]")
+
+    if issyntaxerror and has_filename and has_len:
+        _render_syntax_error(exc, heading=False)
+
+    if isinstance(exc, NameError):
+        suggest_name_error(exc, last_tb)
+
+    if isinstance(exc, AttributeError):
+        suggest_attribute_error(exc)
+
+    if isinstance(exc, ModuleNotFoundError):
+        suggest_module_not_found_error(exc)
+
+    if isinstance(exc, ImportError):
+        suggest_import_error(exc)
+
+    _print_notes(exc)
 
 
 def _print_verbose(
@@ -257,9 +460,9 @@ def _print_verbose(
     tb: TracebackType,
 ) -> None:
     if not _has_rich:
-        print(f"{title}".center(50, "-") if title is not None else "")
+        print(f"{title}".center(50, "─") if title is not None else "")
     else:
-        print(f"[red]{title}[/red]".center(50, "-") if title is not None else "")
+        print(f"[red]{title}[/red]".center(50, "─") if title is not None else "")
     prev_key = None
     count = 0
     prev_frame = None
@@ -291,7 +494,7 @@ def _print_verbose(
                         print(
                             f"[cyan](Previous line repeated {count-1} more times)[/cyan]"
                         )
-                print("-" * 40)
+                print("─" * 40)
             prev_frame = frame
             prev_key = key
             count = 1
@@ -316,40 +519,17 @@ def _print_verbose(
                 _show_context(prev_frame.filename, prev_frame.lineno)
             if count > 3:
                 print(f"[cyan](Previous line repeated {count-1} more times)[/cyan]")
-        print("-" * 40)
+        print("─" * 40)
 
     while tb.tb_next:
         tb = tb.tb_next
 
-    filtered = []
-    for k, v in tb.tb_frame.f_locals.items():
-        if k.lower() in [
-            "password",
-            "token",
-            "key",
-            "api_key",
-            "api_token",
-            "api_password",
-        ]:
-            continue
-        if k.startswith("__"):
-            continue
-        if inspect.ismodule(v):
-            continue
-        if inspect.isfunction(v):
-            continue
-        if isinstance(v, BaseException):
-            continue
-        try:
-            val = _safe_repr.repr(v)
-        except Exception:
-            val = "<repr broken>"
-        filtered.append((k, val, type(v).__name__))
+    filtered = _give_filtered_locals(tb)
     if config.show_locals and filtered:
         if not _has_rich:
-            print("\nLocal variables (last frame):")
+            print("Local variables (last frame):")
         else:
-            print("[yellow]\nLocal variables (last frame):[/yellow]")
+            print("[yellow]Local variables (last frame):[/yellow]")
         for k, v, t in filtered:
             print(f" {k} ({t}) = {v}")
 
@@ -401,9 +581,9 @@ def _print_context(
     count = 0
     prev_frame = None
     if not _has_rich:
-        print("-- Traceback (context mode) --")
+        print("── Traceback (context mode) ──")
     else:
-        print("-- [red]Traceback (context mode)[/red] --")
+        print("── [red]Traceback (context mode)[/red] ──")
     for frame in frames:
         key = (frame.filename, frame.name)
         if key == prev_key:
@@ -432,7 +612,7 @@ def _print_context(
                         print(
                             f"[cyan](Previous line repeated {count-1} more times)[/cyan]"
                         )
-                print("-" * 40)
+                print("─" * 40)
             prev_frame = frame
             prev_key = key
             count = 1
@@ -457,7 +637,7 @@ def _print_context(
                 _show_context(prev_frame.filename, prev_frame.lineno, 1)
             if count > 3:
                 print(f"[cyan](Previous line repeated {count-1} more times)[/cyan]")
-        print("-" * 40)
+        print("─" * 40)
     while tb.tb_next:
         tb = tb.tb_next
 
@@ -506,9 +686,9 @@ def _print_compact(
 ) -> None:
     frames = frames[-3:]
     if not _has_rich:
-        print("-- Traceback (compact mode) --")
+        print("── Traceback (compact mode) ──")
     else:
-        print("[red]-- Traceback (compact mode) --[/red]")
+        print("[red]── Traceback (compact mode) ──[/red]")
     for i, frame in enumerate(frames):
         is_last = i == len(frames) - 1
 
@@ -562,7 +742,7 @@ def _print_minimal(
 ) -> None:
     _hidden_count = len(traceback.extract_tb(tb)) - 1
 
-    print("[red]-- Exception --[/red]") if _has_rich else print("-- Exception --")
+    print("[red]── Exception ──[/red]") if _has_rich else print("── Exception ──")
     frames: traceback.StackSummary = frames[-1:]
     for frame in frames:
         if not _has_rich:
@@ -584,6 +764,7 @@ def _print_minimal(
 
 
 _MODES = {
+    "debug": _print_debug,
     "verbose": _print_verbose,
     "context": _print_context,
     "compact": _print_compact,
@@ -634,6 +815,8 @@ def _print_tb(
         return
     if config.mode == "verbose":
         _MODES["verbose"](title, frames, exc_type, exc, tb)
+    elif config.mode == "debug":
+        _MODES["debug"](title, frames, exc_type, exc, tb)
     else:
         _MODES[config.mode](frames, exc_type, exc, tb)
 
@@ -667,13 +850,13 @@ def _customhook(
     try:
         if exc_type and issubclass(exc_type, KeyboardInterrupt):
             if not _has_rich:
-                print(f"Keyboard Interrupt".center(50, "-"))
+                print(f"Keyboard Interrupt".center(50, "─"))
                 print("The program was terminated by the user")
                 print(
                     "Note: If you triggered it accidentally, note that Ctrl + C mean KeyboardInterrupt"
                 )
             else:
-                print("[yellow]Keyboard Interrupt[/yellow]".center(50, "-"))
+                print("[yellow]Keyboard Interrupt[/yellow]".center(50, "─"))
                 print("[yellow]The program was terminated by the user[/yellow]")
                 print(
                     "[cyan]Note: [/cyan]If you triggered it accidentally, note that Ctrl + C means KeyboardInterrupt"
@@ -690,11 +873,11 @@ def _customhook(
             _print_tb("An error occurred", type(cause), cause, cause.__traceback__)
             if not _has_rich:
                 print(
-                    "\n----The above exception was the cause of the other exception below----\n"
+                    "\n────The above exception was the cause of the other exception below────\n"
                 )
             else:
                 print(
-                    "\n----[red]The above exception was the cause of the other exception below[/red]----\n"
+                    "\n────[red]The above exception was the cause of the other exception below[/red]────\n"
                 )
             _print_tb("Another error occurred", exc_type, exc, tb)
         elif exc and exc.__context__ and not exc.__suppress_context__:
@@ -702,11 +885,11 @@ def _customhook(
             _print_tb("An error occurred", type(ctx), ctx, ctx.__traceback__)
             if not _has_rich:
                 print(
-                    "\n----While handling the previous exception, a new exception has occurred----\n"
+                    "\n────While handling the previous exception, a new exception has occurred────\n"
                 )
             else:
                 print(
-                    "\n----[red]While handling the previous exception, a new exception has occurred[/red]----\n"
+                    "\n────[red]While handling the previous exception, a new exception has occurred[/red]────\n"
                 )
             _print_tb("Another error occurred", exc_type, exc, tb)
         else: