+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com)  
 and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [Unreleased]
+
+## Added
+- A new `debug` mode
+
+## Changed
+- Traceback formatting
+
 ## [0.3.0]  - 2026-06-17
 
 ### Added